backend/requirement.txt
fastapi
uvicorn
python-dotenv
requests
openai
pypdf
python-docx
supabase
cd 